+++ src/lib.rs
@@ -45,7 +45,7 @@ pub mod internal;
 //-----------------------------------------------------------------------------
 
 pub use crate::bwt::Pos;
-pub use crate::gbwt::{GBWT, SearchState, BidirectionalState, Metadata, PathName};
+pub use crate::gbwt::{GBWT, SearchState, BidirectionalState, Metadata, PathName, FullPathName};
 pub use crate::gbz::GBZ;
 pub use crate::graph::Segment;
 pub use crate::support::Orientation;